fix rebase + undo state machine changes
